pyproject.toml
[project]
name = "crawlagent"
version = "1.0.0"
description = "LangGraph Multi-Agent Self-Healing Web Crawler"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "scrapy>=2.11.0",
    "trafilatura>=1.12.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=1.0.2",
    "langchain-community>=0.2.0",
    "google-generativeai>=0.8.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pytest>=7.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
    "apscheduler (>=3.11.1,<4.0.0)",
    "matplotlib (>=3.5.0)",
    "pillow (>=9.0,<12.0)",
    "tavily-python>=0.5.0",
    "firecrawl-py>=1.0.0",
    "psutil (>=7.1.3,<8.0.0)",
    "html2text (>=2025.4.15,<2026.0.0)",
    "markdownify (>=1.2.0,<2.0.0)",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v -s"
markers = [
    "integration: marks tests as integration tests (API calls, external dependencies)",
    "slow: marks tests as slow running tests (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

# [dependency-groups]
# LangGraph CLI는 LangGraph Studio Desktop용이므로 제거
# LangSmith를 사용하므로 불필요함

[dependency-groups]
dev = [
    "langgraph-cli[inmem] (>=0.4.7,<0.5.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pylint (>=4.0.3,<5.0.0)",
    "bandit (>=1.8.6,<2.0.0)",
    "safety (>=3.7.0,<4.0.0)"
]
